refactor: rename http to interfaces
