+++ src/domains/tools/factories/controllers/create-tool-controller.factory.ts
@@ -1,7 +1,7 @@
 import { buildCreateToolValidation } from "@/domains/tools/validation/implementations/zod";
 import { PrismaToolRepository } from "@/domains/tools/repositories/implementations/prisma";
 import { createToolUseCase } from "@/domains/tools/usecases/create-tool.usecase";
-import { createToolController } from "@/domains/tools/http/controllers";
+import { createToolController } from "@/domains/tools/interfaces/controllers";
 
 export const createToolControllerFactory = () => {
   const prismaToolRepository = new PrismaToolRepository();
