+++ src/domains/tools/factories/controllers/list-tools-controller.factory.ts
@@ -1,6 +1,6 @@
 import { listToolUseCase } from "@/domains/tools/usecases/list-tools.usecase";
 import { buildListToolValidation } from "@/domains/tools/validation/implementations/zod";
-import { listToolController } from "@/domains/tools/http/controllers";
+import { listToolController } from "@/domains/tools/interfaces/controllers";
 import { PrismaToolRepository } from "@/domains/tools/repositories/implementations/prisma";
 
 export const listToolControllerFactory = () => {
