+++ src/domains/tools/factories/resolvers/delete-tool-resolver.factory.ts
@@ -1,4 +1,4 @@
-import { deleteToolResolver } from "@/domains/tools/http/resolvers";
+import { deleteToolResolver } from "@/domains/tools/interfaces/resolvers";
 import { deleteToolUseCase } from "@/domains/tools/usecases/delete-tool.usecase";
 import { PrismaToolRepository } from "@/domains/tools/repositories/implementations/prisma";
 import { buildDeleteToolValidation } from "@/domains/tools/validation/implementations/zod";
