+++ src/domains/tools/factories/resolvers/list-tools-resolver.factory.ts
@@ -1,5 +1,5 @@
 import { listToolUseCase } from "@/domains/tools/usecases";
-import { listToolResolver } from "@/domains/tools/http/resolvers";
+import { listToolResolver } from "@/domains/tools/interfaces/resolvers";
 import { buildListToolValidation } from "@/domains/tools/validation/implementations/zod";
 import { PrismaToolRepository } from "@/domains/tools/repositories/implementations/prisma";
