+++ src/domains/tools/interfaces/controllers/delete-tool.controller.spec.ts
