+++ src/domains/tools/interfaces/resolvers/delete-tool.resolver.ts
