+++ src/domains/tools/interfaces/resolvers/index.ts
