+++ src/domains/tools/interfaces/resolvers/list-tools.resolver.ts
