+++ src/domains/tools/interfaces/routes/tool.ts
